% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity-package.R
\docType{package}
\name{multiplicity-package}
\alias{multiplicity-package}
\title{multiplicity: transforming purity ploidy}
\description{
This package is responsible for the total transformation of counts to total copies based on the estimated CN, inferred purity, ploidy discerned by JaBbA.
}
\author{
\strong{Maintainer}: Johnathan Rafailov \email{jrafailov12@gmail.com}

}
