% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{multiplicity}
\alias{multiplicity}
\title{Converts counts to copies.}
\usage{
multiplicity(
  somatic_snv = NULL,
  germline_snv = NULL,
  het_pileups_wgs = NULL,
  tumor_cbs = NULL,
  tumor_dryclean = NULL,
  dryclean_field,
  jabba_rds,
  mask = NULL,
  inferred_sex = NA,
  read_size = 151,
  snpeff_path,
  tumor_name = NULL,
  normal_name = NULL,
  filterpass = FALSE,
  tau_in_gamma = FALSE,
  purity = NULL,
  ploidy = NULL,
  verbose = FALSE
)
}
\arguments{
\item{somatic_snv}{Path to somatic SNV file.}

\item{germline_snv}{Path to germline SNV file.}

\item{het_pileups_wgs}{Path to het_pileups_wgs file.}

\item{tumor_cbs}{Path to segmented drycleaned coverage file. will override supplied tumor_dryclean file}

\item{tumor_dryclean}{Path to drycleaned coverage file. if provided, will rescale REF/ALT values to expected base coverage as determined by dryclean.}

\item{dryclean_field}{specification of field in tumor_dryclean gRanges used as rescaled binned coverage.}

\item{jabba_rds}{Path to jabba file}

\item{mask}{if mask gRanges is provided, loci within ranges will be excluded from analysis}

\item{inferred_sex}{enum of \link{"M", "F"} that indicates inferred of true sex of patient sample (default = NA); optional input; will be inferred from coverage/JaBbA graph if not provided}

\item{read_size}{estimated average read size (default: 151 for Illumina sequencers.)}

\item{snpeff_path}{Path to unzipped SnpEff toolkit}

\item{tumor_name}{Expected name of tumor as annotated in the VCF; highly suggested this is provided for accurate VCF parsing}

\item{normal_name}{Expected name of normal as annotated in the VCF; highly suggested this is provided for accurate VCF parsing}

\item{filterpass}{process only FILTER == PASS variants?}

\item{tau_in_gamma}{Use tau(T) or tau_hat(F) in computation of gamma? tau_hat is the average copy number only of loci; tau is simply ploidy}

\item{purity}{Purity of inputted jabba_rds (optional if metadata of gGraph contains purity)}

\item{ploidy}{Ploidy of inputted jabba_rds (optional if metadata of gGraph contains ploidy)}

\item{verbose}{verbose output?}
}
\value{
Returns a GRanges with counts and converted copies
}
\description{
Converts counts to copies.
}
