% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/multiplicity.R
=======
% Please edit documentation in R/snvplicity.R
>>>>>>> 91df76bf9c62122e5a84b09139bcef71a00b7a79
\name{normalize_path}
\alias{normalize_path}
\title{easy function that returns NULL if file path set to /dev/null}
\usage{
normalize_path(path)
}
\description{
easy function that returns NULL if file path set to /dev/null
}
