% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{parsesnpeff}
\alias{parsesnpeff}
\title{Parse SnpEff vcfs.}
\usage{
parsesnpeff(
  vcf,
  snpeff_path = system.file("extdata", "snpeff_scripts", package = "multiplicity"),
  tumor_id = NULL,
  normal_id = NULL,
  filterpass = TRUE,
  coding_alt_only = TRUE,
  geno = NULL,
  gr = NULL,
  keepfile = FALSE,
  altpipe = FALSE,
  debug = FALSE,
  verbose = FALSE,
  bcftools = "bcftools"
)
}
\arguments{
\item{vcf}{path to snpeff vcf}

\item{snpeff_path}{Path to unzipped SnpEff toolkit}

\item{tumor_id}{Tumor name as annotated in the VCF}

\item{normal_id}{Normal name as annotated in the VCF}

\item{filterpass}{lorem impsum}

\item{coding_alt_only}{lorem impsum}

\item{geno}{lorem impsum}

\item{gr}{lorem impsum}

\item{keepfile}{todo}

\item{altpipe}{lorem impsum}

\item{debug}{lorem impsum}

\item{verbose}{verbose output?}

\item{bcftools}{Path to bcftools toolkit}

\item{pad}{Exposed argument to skitools::ra.overlaps()}

\item{kepefile}{lorem impsum}
}
\value{
GRangesList of breakpoint pairs with junctions that overlap removed

GRangesList of breakpoint pairs with junctions that overlap removed
}
\description{
Ingests and processes a snpeff annotated vcf. #' VCF annotations from a snpeff file are parsed,  with optional functionality provided to select protein coding only mutations or filter = PASS events. AD and DP fields are also collected for downstream applications.
}
