% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_multiplicity_gtrack}
\alias{plot_multiplicity_gtrack}
\title{Create a gTrack for Multiplicity Results}
\usage{
plot_multiplicity_gtrack(
  multiplicity_gr,
  name = "Multiplicity",
  y_fields = c("total_snv_copies", "major_snv_copies", "minor_snv_copies"),
  ...
)
}
\arguments{
\item{multiplicity_gr}{A GRanges object from the output of \code{multiplicity()}.}

\item{name}{The name for the track}

\item{y_fields}{Fields to plot on the y-axis}

\item{...}{Additional arguments passed to gTrack}
}
\value{
A gTrack object
}
\description{
Generates a gTrack object to visualize multiplicity results along the genome,
which can be plotted alongside other genomic data.
}
