% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_multiplicity_histogram}
\alias{plot_multiplicity_histogram}
\title{Plot Multiplicity Histograms}
\usage{
plot_multiplicity_histogram(
  multiplicity_gr,
  field = "total_snv_copies",
  facet_by_cn = TRUE,
  title = "SNV Copy Number Distribution",
  max_cn = 10,
  binwidth = 0.1
)
}
\arguments{
\item{multiplicity_gr}{A GRanges object from the output of \code{multiplicity()}.}

\item{field}{The field to plot, e.g., "total_snv_copies", "major_snv_copies", "minor_snv_copies"}

\item{facet_by_cn}{Logical, whether to facet the histograms by integer copy number}

\item{title}{The title for the plot}

\item{max_cn}{Maximum copy number to include in the plot}

\item{binwidth}{Width of histogram bins}
}
\value{
A ggplot object
}
\description{
Creates histograms to visualize the distribution of estimated SNV copy numbers
from the output of the \code{multiplicity()} function.
}
