% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_vaf_cn_scatter}
\alias{plot_vaf_cn_scatter}
\title{Plot VAF vs. SNV Copy Number}
\usage{
plot_vaf_cn_scatter(
  multiplicity_gr,
  plot_type = c("altered_copies", "altered_fraction"),
  color_by_cn = TRUE,
  facet_by_cn = TRUE,
  title = NULL,
  max_cn = 10
)
}
\arguments{
\item{multiplicity_gr}{A GRanges object from the output of \code{multiplicity()}.}

\item{plot_type}{Either "altered_copies" or "altered_fraction" to determine what to plot on y-axis}

\item{color_by_cn}{Logical, whether to color points by integer copy number}

\item{facet_by_cn}{Logical, whether to facet the plot by integer copy number}

\item{title}{The title for the plot (if NULL, a default title based on plot_type will be used)}

\item{max_cn}{Maximum copy number to include in the plot}
}
\value{
A ggplot object
}
\description{
Creates a scatter plot to visualize the relationship between Variant Allele
Frequency (VAF) and estimated SNV copy numbers or altered fraction.
}
