% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{preprocess_multiplicity_inputs}
\alias{preprocess_multiplicity_inputs}
\title{Preprocess inputs for multiplicity}
\usage{
preprocess_multiplicity_inputs(
  somatic_snv,
  germline_snv,
  het_pileups_wgs,
  tumor_cbs,
  tumor_dryclean,
  dryclean_field,
  jabba_rds,
  mask,
  inferred_sex,
  read_size,
  ploidy,
  purity,
  binwidth = NULL,
  verbose
)
}
\arguments{
\item{somatic_snv}{Path to somatic SNV file.}

\item{germline_snv}{Path to germline SNV file.}

\item{het_pileups_wgs}{Path to het_pileups_wgs file.}

\item{tumor_cbs}{Path to segmented drycleaned coverage file.}

\item{tumor_dryclean}{Path to drycleaned coverage file.}

\item{dryclean_field}{Specification of field in tumor_dryclean gRanges.}

\item{jabba_rds}{Path to jabba file or gGraph object.}

\item{mask}{Path to mask gRanges RDS file or gRanges object.}

\item{inferred_sex}{Enum of \link{"M", "F"}.}

\item{read_size}{Estimated average read size.}

\item{ploidy}{Ploidy of the sample.}

\item{purity}{Purity of the sample.}

\item{verbose}{Verbose output?}
}
\value{
A list containing preprocessed inputs:
somatic_snv, germline_snv, het_pileups_wgs, cn.gr,
dryclean.cov, inferred_sex, purity, ploidy, mask.
}
\description{
Helper function to preprocess and validate inputs for the main multiplicity function.
}
\keyword{internal}
