% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvplicity.R
\name{transform_hets}
\alias{transform_hets}
\title{function to inhale SNV-laden VCF and transform counts to to estimated altered copies using linear transformation via beta, gamma conversions}
\usage{
transform_hets(
  hets,
  cn_gr,
  dryclean.cov = NULL,
  basecov_field = "avg_basecov",
  ploidy,
  purity,
  inferred_sex,
  verbose = TRUE,
  tau_in_gamma = FALSE,
  mask = mask,
  normfactor = 2
)
}
\arguments{
\item{dryclean.cov}{Tumor drycleaned coverage for denoising/rescaling of REF/ALT counts (OPTIONAL; if provided will denoise)}

\item{basecov_field}{field name of dryclean.cov that indicates average Base Coverage per bin}

\item{ploidy}{Inferred ploidy input for conversion}

\item{purity}{Inferred purity input for conversion}

\item{inferred_sex}{enum of \link{"M", "F"} that indicates inferred of true sex of patient sample}

\item{vcf}{Path to snpeff vcf}

\item{tumor_id}{Tumor name as annotated in the VCF (highly suggested this is provided)}

\item{normal_id}{Normal name as annotated in the VCF (highly suggested this is provided)}

\item{filterpass}{process only FILTER == PASS variants?}

\item{snpeff_path}{Path to unzipped SnpEff toolkit}
}
\value{
GRanges of transformed read counts
}
\description{
function to inhale SNV-laden VCF and transform counts to to estimated altered copies using linear transformation via beta, gamma conversions
}
