% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{transform_snv}
\alias{transform_snv}
\title{function to inhale SNV-laden VCF and transform counts to to estimated altered copies using linear transformation via beta, gamma conversions}
\usage{
transform_snv(
  vcf,
  cn_gr,
  dryclean.cov = NULL,
  basecov_field = "avg_basecov",
  major_gamma_coeff = 1,
  minor_gamma_coeff = 1,
  tumor_id = NA,
  normal_id = NA,
  ploidy,
  purity,
  inferred_sex,
  filterpass = FALSE,
  snpeff_path,
  verbose = TRUE,
  tau_in_gamma = FALSE,
  mask = mask
)
}
\arguments{
\item{vcf}{Path to SnpEff-annotated VCF}

\item{dryclean.cov}{Tumor drycleaned coverage for denoising/rescaling of REF/ALT counts (OPTIONAL; if provided will denoise)}

\item{basecov_field}{field name of dryclean.cov that indicates average Base Coverage per bin}

\item{tumor_id}{Tumor name as annotated in the VCF (highly suggested this is provided)}

\item{normal_id}{Normal name as annotated in the VCF (highly suggested this is provided)}

\item{ploidy}{Inferred ploidy input for conversion}

\item{purity}{Inferred purity input for conversion}

\item{inferred_sex}{enum of \link{"M", "F"} that indicates inferred of true sex of patient sample}

\item{filterpass}{process only FILTER == PASS variants?}

\item{snpeff_path}{Path to unzipped SnpEff toolkit}

\item{verbose}{verbose output?}

\item{tau_in_gamma}{Use tau(T) or tau_hat(F) in computation of gamma? tau_hat is the average copy number only of loci; tau is simply ploidy}

\item{mask}{if mask gRanges is provided, loci within ranges will be excluded from analysis}
}
\value{
GRanges of transformed read counts
}
\description{
function to inhale SNV-laden VCF and transform counts to to estimated altered copies using linear transformation via beta, gamma conversions
}
