% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{try2}
\alias{try2}
\title{wrapper around tryCatch - robust to parallel:: functions; A slightly more robust version of try that works within the parallel:: set of functions that pre-deploy a cluster.}
\usage{
try2(expr, ..., finally)
}
\description{
wrapper around tryCatch - robust to parallel:: functions; A slightly more robust version of try that works within the parallel:: set of functions that pre-deploy a cluster.
}
