% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{normalize_path}
\alias{normalize_path}
\title{easy function that returns NULL if file path set to /dev/null}
\usage{
normalize_path(path)
}
\description{
easy function that returns NULL if file path set to /dev/null
}
