% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvplicity.R
\name{parsesnpeff}
\alias{parsesnpeff}
\title{Parse SnpEff vcfs.}
\usage{
parsesnpeff(
  vcf,
  snpeff_path = "~/modules/SnpEff/source/snpEff",
  tumor_id = NULL,
  normal_id = NULL,
  filterpass = TRUE,
  coding_alt_only = TRUE,
  geno = NULL,
  gr = NULL,
  keepfile = FALSE,
  altpipe = FALSE,
  debug = FALSE,
  verbose = FALSE,
  bcftools = "bcftools"
)
}
\arguments{
\item{vcf}{path to snpeff vcf}

\item{snpeff_path}{Path to unzipped SnpEff toolkit}

\item{tumor_id}{Tumor name as annotated in the VCF}

\item{normal_id}{Normal name as annotated in the VCF}

\item{filterpass}{process only FILTER == PASS variants?}

\item{coding_alt_only}{process only coding alterations?}

\item{geno}{todo}

\item{gr}{todo}

\item{keepfile}{todo}

\item{altpipe}{parsesnpeff via khtools::grok_vcf?}

\item{debug}{debug function?}

\item{verbose}{verbose output?}

\item{bcftools}{Path to bcftools toolkit}
}
\value{
GRangesList of breakpoint pairs with junctions that overlap removed
}
\description{
Ingests and processes a snpeff annotated vcf. #' VCF annotations from a snpeff file are parsed,  with optional functionality provided to select protein coding only mutations or filter = PASS events. AD and DP fields are also collected for downstream applications.
}
