% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvplicity.R
\name{parsesnpeff}
\alias{parsesnpeff}
\title{how is this not a paragraph?}
\usage{
parsesnpeff(
  vcf,
  snpeff_path,
  tumor_id = NULL,
  normal_id = NULL,
  filterpass = TRUE,
  coding_alt_only = TRUE,
  geno = NULL,
  gr = NULL,
  keepfile = FALSE,
  altpipe = FALSE,
  debug = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{vcf}{path to snpeff vcf}

\item{tumor_id}{Tumor name as annotated in the VCF}

\item{normal_id}{Normal name as annotated in the VCF}

\item{filterpass}{lorem impsum}

\item{coding_alt_only}{lorem impsum}

\item{geno}{lorem impsum}

\item{gr}{lorem impsum}

\item{altpipe}{lorem impsum}

\item{debug}{lorem impsum}

\item{pad}{Exposed argument to skitools::ra.overlaps()}

\item{kepefile}{lorem impsum}
}
\value{
GRangesList of breakpoint pairs with junctions that overlap removed
}
\description{
how is this not a paragraph?
}
