% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvplicity.R
\name{snvplicity}
\alias{snvplicity}
\title{Predecated name for multiplicity}
\usage{
snvplicity(...)
}
\description{
predecated name for power function to run multiplicity
}
\details{
Please use multiplicity() instead
}
