% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvplicity.R
\name{snvplicity}
\alias{snvplicity}
\title{Converts counts to copies.}
\usage{
snvplicity(
  somatic_snv,
  germline_snv,
  jabba_rds,
  snpeff_path,
  tumor_name = NULL,
  normal_name = NULL
)
}
\arguments{
\item{somatic_snv}{Path to somatic SNV file}

\item{germline_snv}{Path to germline SNV file}

\item{jabba_rds}{Path to jabba file}

\item{snpeff_path}{Path to unzipped SnpEff toolkit}

\item{tumor_name}{Expected name of tumor as annotated in the VCF}

\item{normal_name}{Expected name of normal as annotated in the VCF}
}
\value{
Returns a GRanges with counts and converted copies
}
\description{
Converts counts to copies.
}
