% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{transform_hets}
\alias{transform_hets}
\title{function to inhale SNP-laden gRanges and transform counts to to estimated altered copies using linear transformation via beta, gamma conversions}
\usage{
transform_hets(
  hets,
  cn_gr,
  dryclean.cov = NULL,
  basecov_field = "avg_basecov",
  ploidy,
  purity,
  inferred_sex,
  verbose = TRUE,
  tau_in_gamma = FALSE,
  mask = mask
)
}
\arguments{
\item{hets}{heterozygous SNPs GRanges}

\item{dryclean.cov}{Tumor drycleaned coverage for denoising/rescaling of REF/ALT counts (OPTIONAL; if provided will denoise)}

\item{basecov_field}{field name of dryclean.cov that indicates average \verb{Base Coverage} per bin}

\item{ploidy}{Inferred ploidy input for conversion}

\item{purity}{Inferred purity input for conversion}

\item{inferred_sex}{enum of \link{"M", "F"} that indicates inferred of true sex of patient sample}

\item{tau_in_gamma}{Use tau(T) or tau_hat(F) in computation of gamma? tau_hat is the average copy number only of loci; tau is simply ploidy}

\item{mask}{if mask gRanges is provided, loci within ranges will be excluded from analysis}

\item{tumor_id}{Tumor name as annotated in the VCF (highly suggested this is provided; otherwise inferences are made about ordering in the VCF)}

\item{normal_id}{Normal name as annotated in the VCF (highly suggested this is provided; otherwise inferences are made about ordering in the VCF)}
}
\value{
GRanges of transformed read counts
}
\description{
function to inhale SNP-laden gRanges and transform counts to to estimated altered copies using linear transformation via beta, gamma conversions
}
